\name{grp_pars_dig_01}
\docType{data}
\alias{inf_pars_dig_01}
\alias{sub_pars_dig_01}
\title{ Default parameters for \code{\link{abmenoe}}. }
\description{
  Default microbial group parameters for the \code{\link{abmneo}} function.
  A single set "dig_01" is presently available, and was fit to methane emission measurements from a set of three digestate storage tanks in Denmark.
  Measurements are described in Kasper et al. (2026) and original measurements were shared by first author P. Kasper.
}

\usage{grp_pars_dig_01}
\format{ All parameter objects are lists. See \code{\link{abmneo}} for required elements. }
\references{
Kasper, P.L., Lycus, P.M. & Holm, M., 2026. Field-scale assessment of ammonia and methane emissions from tent-covered slurry storages using high-resolution tracer decay measurements. Biosystems Engineering, 270, p. 104551. \doi{hj10.3168/jds.2012-5340}.
}

\keyword{datasets}
